move all version split functionality to subgait class
