+++ march_shared_classes/src/march_shared_classes/gait/gait.py
@@ -1,5 +1,4 @@
 import os
-import re
 
 import yaml
 
@@ -8,8 +7,6 @@ from march_shared_classes.exceptions.general_exceptions import FileNotFoundError
 
 from .subgait import Subgait
 from .subgait_graph import SubgaitGraph
-
-PARAMETRIC_GATIS_CHARACTER = '_'
 
 
 class Gait(object):
@@ -89,18 +86,7 @@ class Gait(object):
             raise SubgaitNameNotFound(subgait_name, gait_name)
 
         version = gait_version_map[gait_name][subgait_name]
-        if version[0] == PARAMETRIC_GATIS_CHARACTER:
-            base_version, other_version, parameter = unpack_parametric_version(version)
-            base_path = os.path.join(gait_directory, gait_name, subgait_name, base_version + '.subgait')
-            other_path = os.path.join(gait_directory, gait_name, subgait_name, other_version + '.subgait')
-
-            return Subgait.from_files_interpolated(robot, base_path, other_path, parameter)
-        else:
-            subgait_path = os.path.join(gait_directory, gait_name, subgait_name, version + '.subgait')
-            if not os.path.isfile(subgait_path):
-                raise FileNotFoundError(file_path=subgait_path)
-
-            return Subgait.from_file(robot, subgait_path)
+        return Subgait.from_name_and_version(robot, gait_directory, gait_name, subgait_name, version)
 
     def _validate_trajectory_transition(self):
         """Compares and validates the trajectory end and start points."""
@@ -121,22 +107,11 @@ class Gait(object):
         :param dict version_map: Mapping subgait names to versions
         """
         new_subgaits = {}
-        gait_path = os.path.join(gait_directory, self.gait_name)
         for subgait_name, version in version_map.items():
             if subgait_name not in self.subgaits:
                 raise SubgaitNameNotFound(subgait_name, self.gait_name)
-            if version[0] == PARAMETRIC_GATIS_CHARACTER:
-                base_version, other_version, parameter = unpack_parametric_version(version)
-                base_version_path = os.path.join(gait_path, subgait_name, base_version + '.subgait')
-                other_version_path = os.path.join(gait_path, subgait_name, other_version + '.subgait')
-                new_subgaits[subgait_name] = Subgait.from_files_interpolated(robot, base_version_path,
-                                                                             other_version_path, parameter)
-
-            else:
-                subgait_path = os.path.join(gait_path, subgait_name, version + '.subgait')
-                if not os.path.isfile(subgait_path):
-                    raise FileNotFoundError(file_path=subgait_path)
-                new_subgaits[subgait_name] = Subgait.from_file(robot, subgait_path)
+            new_subgaits[subgait_name] = Subgait.from_name_and_version(robot, gait_directory, self.gait_name,
+                                                                       subgait_name, version)
 
         for from_subgait_name, to_subgait_name in self.graph:
             if from_subgait_name in new_subgaits or to_subgait_name in new_subgaits:
@@ -154,12 +129,3 @@ class Gait(object):
     def __getitem__(self, name):
         """Returns a subgait from the loaded subgaits."""
         return self.subgaits.get(name)
-
-
-def unpack_parametric_version(version):
-    parameter_str = re.search(r'{0}[0-9.]*_'.format(PARAMETRIC_GATIS_CHARACTER), version).group(0)
-    parameter = float(parameter_str[1:-1])
-    versions = re.findall(r'\([^\)]*\)', version)
-    base_version = versions[0][1:-1]
-    other_version = versions[1][1:-1]
-    return base_version, other_version, parameter
